property_valuation/requirements.txt
# Core
numpy>=1.23
pandas>=1.5
scipy>=1.10

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Machine Learning
scikit-learn>=1.2
xgboost>=1.7
joblib>=1.2

# Deep Learning
torch>=2.0
torchvision>=0.15

# Image Processing
opencv-python>=4.8
Pillow>=9.5

# Utilities
tqdm>=4.65
pyyaml>=6.0

# Jupyter / Notebooks
jupyter
notebook
